+++ changelog.md
@@ -33,6 +33,7 @@ From colorful bell peppers to cozy cups of coffee, Rustic Delight 1.7 is packed 
 
 ### Changes
 - Updated Farmer's Delight Refabricated to 3.6.5 (now the minimum required version)
+- Updated Japanese translations (thank you HayaKoh!)
 - You no longer need a knife to get a single pancake: Pancakes are now taken from the block by right-clicking, instead of being eaten directly.
 - Breaking a pancake block now drops the remaining pancakes as individual pancake items
 - Updated Rice Roll Royale texture & models to match Farmer's Delight latest Rice Roll Medley changes